(#158) Add CLI ordering to @req_metadata_hash in Model classes
In Ruby 1.8.7, Hash traveral is not deterministic unlike Ruby 1.9 (which
is insertion order based). This leads to a situation where the Model
metdata CLI questions are asked in an essentially random order.

An order/rank/position attribute in each of the key's subhashes will
provide a way to conistently ask Model-related questions in the same
order across different Ruby versions.