Batch Mode (REPL)
This provides the ability to start the razor CLI up in a long running
process and issue commands to it via stdin. Each line sent to stdin will
be read and parsed as if it were sent via the shell.

Combining --batch with the --webcommand option allows us to start up a
pool of long running razor commands to process incoming requests. This
should improve razor's performance when under heavy load.