Image Service IO Pool
When a large number of requests is received the Image Service can
exhaust the IO of potential of the system. Without some kind of
throttling, it can spawn a large number of simulaneous IO requests which
then cause resource contention issues to drive overall IO performance
down.

The IO pool is limited to 5 simultaneous IO streams. When the systems IO
is not saturated this has very little impact on the overall IO
performance. When the system is IO constrained it helps keep the
existing IO operations progressing.