If puppet auto-signing is off, avoid that MK get booted always
The broker_wait (aka. system_wait) state is today used only for puppet broker.
The state gets entered on puppetmaster setups with no auto-siging.
There is no magic to switch from this state to broker_success.
Since the broker_wait is not covered in the model mk_call or boot_call this
caused that any installation after the broker handover ended up in
bootin Razor MK again - since the active model was still in state
broker_wait after the handover.