Puppet Broker modification for DTAG cloud
We go with our intenral puppet packages from our internal mirror.
This needs to be made configurable to make it upstream acceptable.

To allow puppet to reconfigure the network we first run puppet with
noop to check if the puppet manifest have no syntax error. If there
would be errors the run_agent() comment would return the stderr
message from that command back to Razor which could then indicate
if the broker handover failed, succeeded or is in a broker_wait
state.

To not let the puppet handover get stuck in postinstall we have
to do the actual puppet agent called daemonize since the network
configuration get changed via puppet and the SSH gets dropped
for some machines.