Fix storage node provisioning with JBOD RAID configuration
With JBOD RAID configuration the JBOD-disks have a lower SCSI bus ID
and get numbered first by kernel/udev before the virtual-disks with
RAID configuration.

Since we will not gurantee that we have always the same number of disks
in future in each storage node we dynamicly detect the virtual-disks
and go with something like /dev/sdz during installation.
Later we stick with dev-by-id configurations.

The sed workaround:

    sed -i 's,db_input critical grub-installer/bootdev,db_input low grub-installer/bootdev,g' /usr/bin/grub-installer

... is required due to an issue with grub-installer in Ubuntu Precise which
breaks the automated installation by presenting the db_input grub-installer/bootdev
dialog. Since this dialog is called with "ciritcal" priority we can not get
rid of that dialog with preseed in a clean fashion.

grub-installer seems to be not deal correctly if the default bootdevice is different
then the one specific via preseed and grub-installer/bootdev.

We could not workaround this either by setting the grub-installer/bootdev seen flag.

fixes SMBCAP-3561